\name{readData}
\alias{readData}
\title{readData}
\usage{
  readData(filename, filetype, verbose = TRUE)
}
\arguments{
  \item{filename}{Filename (indlucing path)}

  \item{filetype}{Custom specified file type}

  \item{verbose}{Verbose output mode}
}
\description{
  Function for reading a specified file into R and saving
  it as a internal dataset
}

